requirements.txt
# Core Django Framework
Django==4.2.7
djangorestframework==3.14.0

# Django Extensions
django-cors-headers==4.3.1
django-filter==23.3
django-storages==1.14.2

# Database
mongoengine==0.27.0
pymongo==4.6.0

# Task Queue
celery==5.3.4
redis==5.0.1

# AWS Services
boto3==1.29.7

# Image Processing
Pillow==11.0.0

# Configuration
python-decouple==3.8

# Authentication & Security
PyJWT==2.8.0
cryptography==41.0.7

# Language Detection
langdetect==1.0.9

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# Data Science & ML
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2

# Development & Testing (Optional - uncomment if needed)
# pytest==7.4.3
# pytest-django==4.7.0
# pytest-cov==4.1.0
# black==23.11.0
# flake8==6.1.0
# mypy==1.7.1
# django-debug-toolbar==4.2.0
# factory-boy==3.3.0